+++ src/jk_argparsing/__init__.py
@@ -1,11 +1,10 @@
 ﻿
 
 
-__version__ = "0.2021.3.11.1"
+__version__ = "0.2021.3.17"
 
 
 
 from .ArgOption import ArgOption
 from .ArgsParser import ArgsParser
 from .ParsedArgs import ParsedArgs
-